requirements.txt
# Streamlit app dependencies
streamlit>=1.30.0
numpy>=1.26.0
pandas>=2.3.0
scikit-learn>=1.7.0
requests>=2.32.0
joblib>=1.5.0

# TensorFlow CPU version compatible with Streamlit
tensorflow-cpu>=2.20.0

# Optional for nicer logs
rich>=14.0
